Improved code to change names of scenarios and use cases for multiple builds in demo data
